get_around_streamlit/requirements.txt
bokeh==3.6.0
numpy==1.26.0
plotly==5.24.1
streamlit-lottie==0.0.5

streamlit==1.37.1
pandas==2.2.1
matplotlib==3.7.3
seaborn==0.13.2
requests==2.32.3
streamlit-option-menu==0.4.0
Pillow==10.3.0
pyarrow==16.1.0
openpyxl==3.1.2
